requirements.txt
# Core retrieval
rank-bm25==0.2.2
sentence-transformers==3.4.1
# Windows: use faiss-cpu. Linux/Mac with NVIDIA GPU: replace with faiss-gpu
faiss-cpu

# API framework
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LLM generation
google-genai>=1.0.0
groq==0.20.0
httpx==0.28.1

# Data
kagglehub==0.3.12
numpy==1.26.4
tqdm==4.67.1

# Utilities
python-dotenv==1.1.0
